+++ src/main/kotlin/com/sletmoe/krogue/algorithms/los/SymmetricShadowCaster.kt
@@ -70,7 +70,7 @@ private class Row(val quadrant: Quadrant, val depth: Int, val startSlope: Fracti
 }
 
 class SymmetricShadowCaster : LineOfSightCalculator {
-    override fun calculateLineOfSight(origin: Point, tiles: Grid<Tile>, maxViewDistance: Int?): Grid<Boolean> {
+    override fun calculateLineOfSight(origin: Point, tiles: Grid<Tile>, maxViewDistance: Double?): Grid<Boolean> {
         val visibilityGrid = Grid(tiles.width, tiles.height, defaultValue = false)
 
         // mark the origin as visible
@@ -133,7 +133,7 @@ class SymmetricShadowCaster : LineOfSightCalculator {
         }
     }
 
-    private fun limitVisibilityDistance(origin: Point, maxVisibilityDistance: Int, visibilityGrid: Grid<Boolean>) {
+    private fun limitVisibilityDistance(origin: Point, maxVisibilityDistance: Double, visibilityGrid: Grid<Boolean>) {
         val visDistanceSquared = maxVisibilityDistance * maxVisibilityDistance
 
         visibilityGrid.forEachCoordinate { coordinate ->
